+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 2025-02-05 deimos 0.6.0
+
+### Added
+
+* Add `DataFrameDispatcher` for writing data to in-memory dataframe
+* Add `Msg::Packet(Vec<u8>)` variant of user channel message to enable generic packetized message-passing
+* !Add `terminate` method to `Dispatcher` and `Calc` traits and `Orchestrator` struct
+    * To be called when the controller terminates, returning dispatchers to their pre-init state for reuse
+
+### Changed
+
+* !Rename `Panic` variant of `Overflow` to `Error`
+    * Dispatchers can return an error on consuming values; this provides more flexibility in future error handling
+* !Rename `initialize` method of `Dispatcher` trait to `init`
+* !Rename `calcs` module to `calc`
+* !Rename `peripherals` module to `peripheral`
+* !Move orchestrator module under `calcs` and export 
+* Make large const arrays static to avoid inlining excessively large data
+
 ## 2025-02-04 deimos 0.5.1
 
 ### Added